requirements.txt
anthropic
google-genai
python-dotenv
rich
